190410| user

+++ django/insta/posts/views.py
@@ -1,31 +1,43 @@
 from django.shortcuts import render,redirect, get_object_or_404
+from django.contrib.auth.decorators import login_required
 from .forms import PostForm
 from .models import Post
 # Create your views here.
 
 def list(request):
-    posts=Post.objects.all()
+    posts=Post.objects.order_by('-id').all()
     return render(request,'posts/list.html',{'posts':posts})
     
+@login_required
 def create(request):
     if request.method=='POST':
         post_form = PostForm(request.POST, request.FILES)
         if post_form.is_valid():
-            post_form.save()
+            post=post_form.save(commit=False)
+            post.user=request.user
+            post_form.save() #실제 데이터베이스에 저장
             return redirect('posts:list')
     else:
         post_form = PostForm()
    
-    return render(request,'posts/create.html',{'post_form':post_form})
+    return render(request,'posts/form.html',{'post_form':post_form})
     
 def delete(request,post_id):
-    post=Post.objects.get(pk=post_id)
-    get_object_or_404(Post, pk=post_id) #pk=id 로 고쳐써도됨
+    # post=Post.objects.get(pk=post_id)
+    
+    post=get_object_or_404(Post, pk=post_id) #pk=id 로 고쳐써도됨
+    if post.user != request.user:
+        return redirect('posts:list')
+    
     post.delete()
     return redirect('posts:list')
     
+    
+    
 def update(request,post_id):
     post = get_object_or_404(Post, id=post_id)
+    if post.user != request.user:
+        return redirect('posts:list')
     if request.method == 'POST':
         post_form = PostForm(request.POST, request.FILES, instance=post)
         if post_form.is_valid():
@@ -34,4 +46,4 @@ def update(request,post_id):
     else:
         post_form = PostForm(instance=post)
     post_form = PostForm(instance=post)
-    return render(request, 'posts/create.html',{'post_form':post_form})
+    return render(request, 'posts/form.html',{'post_form':post_form})
